Ordenar funciones auxiliares en formatFunctions.js
