+++ src/components/admin/ProductTable.jsx
@@ -5,6 +5,7 @@ import Pagination from "./Pagination";
 import Form from "./Form";
 import Modal from "./Modal";
 import Message from "./Message";
+import { priceRangeCalculator, roundToNearest50 } from "../../utils/formatFunctions";
 
 
 const ProductTable = () => {
@@ -59,39 +60,9 @@ const ProductTable = () => {
 		}
 	}, [successMessage, errorMessage]);
 
-	const priceRangeCalculator = (precioRenta) => {
-		const precios = state.data
-			.filter((obra) => obra.precioRenta)
-			.map((obra) => obra.precioRenta);
+	
 
-		if (precios.length === 0) {
-			return "Precio no disponible";
-		}
-
-		const minPrice = Math.min(...precios);
-		const maxPrice = Math.max(...precios);
-		const avgPrice = (minPrice + maxPrice) / 2;
-
-		if (precioRenta <= minPrice + (maxPrice - minPrice) * 0.3) {
-			return "$"; // Precio bajo
-		} else if (precioRenta <= minPrice + (maxPrice - minPrice) * 0.7) {
-			return "$$"; // Precio medio
-		} else {
-			return "$$$"; // Precio alto
-		}
-	};
-
-	const roundToNearest50 = (year) => {
-		// Se calcula el múltiplo de 50 más cercano al año
-		const remainder = year % 50;
-		if (remainder < 25) {
-			// Si el resto es menor que 25, redondeamos hacia abajo
-			return year - remainder;
-		} else {
-			// Si el resto es mayor o igual que 25, redondeamos hacia arriba
-			return year + (50 - remainder);
-		}
-	};
+	
 
 	return (
 		<div className="flex flex-col items-center grow max-h-screen pt-28 relative ">
